fix: Resolve MSVC warnings C4244, C4456 and add #ifndef guards for IMAGE_* macros
- coffi_relocation.hpp: Apply narrow_cast to uint32_t -> uint16_t
  assignments in relocation::save() for TI and default architectures
- coffi.hpp: Rename shadowed local variable 'machines' to
  'ceva_machines' in CEVA architecture detection (C4456)
- coffi_types.hpp: Add #ifndef guards around all IMAGE_* macro
  definitions to prevent C4005 when coffi_types.hpp is included
  before winnt.h
- coffi_utils.hpp: Centralize narrow_cast definition as an inline
  constexpr template function, removing duplicate macro definitions
  from coffi.hpp, coffi_directory.hpp, coffi_headers.hpp,
  coffi_section.hpp, coffi_symbols.hpp

+++ coffi/coffi_relocation.hpp
@@ -119,7 +119,7 @@ class relocation
             rel_entry_ti h;
             h.virtual_address    = header.virtual_address;
             h.symbol_table_index = header.symbol_table_index;
-            h.type               = header.type;
+            h.type               = narrow_cast<uint16_t>(header.type);
             h.reserved           = header.reserved;
             stream.write((char*)&(h), sizeof(h));
             break;
@@ -138,7 +138,7 @@ class relocation
             rel_entry h;
             h.virtual_address    = header.virtual_address;
             h.symbol_table_index = header.symbol_table_index;
-            h.type               = header.type;
+            h.type               = narrow_cast<uint16_t>(header.type);
             stream.write((char*)&(h), sizeof(h));
             break;
         }
